Started AI
I must make it so that the corresponding value in the played array is set to true, and when the corresponding button is clicked again, it does not input, and turn does not increment. Therefore, AI checking will also need to have a condition that checks if the button to be played in is empty. If so, then only can it play in that box.